fix ghosting, implement minimum hover duration
